add calendar model
