+++ .eslintignore
@@ -1,0 +1,2 @@
+/public/javascripts
+/vendor/assets/javascripts
