+++ spec/factories/calendars.rb
@@ -1,0 +1,5 @@
+FactoryGirl.define do
+  factory :calendar do
+    
+  end
+end
